chore: rename find_references to get_references
